+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.8
+
+- Fixes [#7](https://github.com/alexnederlof/Jasper-report-maven-plugin/issues/7) For each `.jrxml` file search th compiled .jasper if not found or older than source file recompile it.
+- Fixes [#9](https://github.com/alexnederlof/Jasper-report-maven-plugin/issues/9) Project classpath added to ClassLoader.
+
+Thanks to [@lucarota](https://github.com/lucarota)
+
 # 1.7
 
  - Allows developers to set additional properties [#5](https://github.com/alexnederlof/Jasper-report-maven-plugin/pull/5). Thanks to [@Plozi](https://github.com/plozi)